Added simulation option
Added an argument to CarPlayManagerDelegate methods indicating the current template. Call these delegate methods for the Favorites List template in addition to the main map template. Added a trailing bar button to the example application that toggles route simulation on and off.

Added a method to CarPlayManagerDelegate for providing the route controller. The example application conditionally provides a route controller with a simulated location manager.