Overhauled navigation map template management
Removed CarPlayMapTemplateController. CarPlayManager is now directly responsible for setting up the navigation map template, starting the navigation session, and responding to gestures on the navigation view controller’s map (with help from the navigation view controller). Consolidated the overview and recentering buttons into a single toggle button.

CarPlayManagerDelegate is now given an opportunity to add navigation bar buttons to the navigation map template alongside the built-in ones. Added an argument to CarPlayManagerDelegate methods to distinguish between the three map templates.